+++ br_account_einvoice/models/__init__.py
@@ -3,7 +3,9 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 
-from . import invoice_electronic
-from . import res_company
-from . import account_invoice
 from . import account_config_settings
+from . import account_invoice
+from . import invoice_electronic
+from . import invoice_electronic_event
+from . import invoice_electronic_item
+from . import res_company
