+++ br_account_einvoice/tests/__init__.py
@@ -2,5 +2,5 @@
 # © 2016 Danimar Ribeiro, Trustcode
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from . import test_electronic_invoice
 from . import test_account_invoice
+from . import test_electronic_invoice
